nrL1_stats: avoid overflow in total_bytes_tx (#471)

I noticed from some nrL1stats.log that total_bytes_tx overflows, for
instance:

    total_bytes TX -2129096305

This PR uses uint64_t instead of int to avoid the overflow.

+++ openair1/PHY/defs_gNB.h
@@ -56,7 +56,7 @@ typedef struct {
 typedef struct {
   int dump_frame;
   int round_trials[8];
-  int total_bytes_tx;
+  uint64_t total_bytes_tx;
   int total_bytes_rx;
   int current_Qm;
   int current_RI;
